fix: password_digestの削除を安全な条件付き実行に変更

+++ db/migrate/20260321042711_remove_password_digest_from_users.rb
@@ -1,5 +1,7 @@
 class RemovePasswordDigestFromUsers < ActiveRecord::Migration[7.2]
   def change
+    if column_exists?(:users, :password_digest)
       remove_column :users, :password_digest, :string
+    end
   end
 end
